feature/ Add a method to split money equally

+++ src/main/java/com/LightSplit/demo/Model/Traveler.java
@@ -24,14 +24,15 @@ public class Traveler {
     @NotNull
     private String nickName;
     
-    private int balance = 0;
+    private double balance = 0;
 
+    // I'm really not confused by what's happening here 
     @Override
     public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
+        if (this == o) return true; // A
+        if (o == null || getClass() != o.getClass()) return false; // 
         Traveler traveler = (Traveler) o;
-        return Objects.equals(id, traveler.id);
+        return Objects.equals(id, traveler.id); // B, I don't get how line A & B need to be seperated; is it because equals method returns error if the compared objects aren't the same class ?
     }
 
     @Override
